requirements.txt
geographiclib==1.49
geopy==1.18.0
httplib2==0.12.0
psutil==5.4.8
pynmea2==1.12.0
pyserial==3.4
simplejson==3.16.0
